# openapi-appengine.yaml
...
swagger: "2.0"
info:
  description: "API for Quran"
  title: "Quran API"
  version: "1.0.0"
host: "utopian-button-227405.appspot.com"
...
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - "https"
paths:
...
  "/v1/ayah/surah/{surahId}":
    get:
      tags:
        - "Ayah"
      description: "Get Ayah by Surah Id"
      operationId: "getAyahBySurahId"
      parameters:
        - in: path
          name: surahId
          type: string
          required: true
          description: String ID of Surah
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
        - in: query
          name: parts
          type: string
          description: Comma seperated additional parts possible value are Translation,Surah,Edition,Arabic_Audio,Translation_Audio,Image. You can set "All" if you want to get all parts.
        - in: query
          name: edition
          type: string
          description: String id of edition
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Ayah"
          schema:
            $ref: "#/definitions/ayah"
...
  "/v1/surah/{surahId}":
    get:
      tags:
        - "Surah"
      description: "Get Surah by ID"
      operationId: "getSurahById"
      parameters:
        - in: path
          name: surahId
          type: string
          required: true
          description: String ID of surah
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Surah"
          schema:
            $ref: "#/definitions/surah"
  "/v1/surah/all":
    get:
      tags:
        - "Surah"
      description: "Get All surah"
      operationId: "getAllSurah"
      parameters:
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Surah"
          schema:
            $ref: "#/definitions/surah"
  "/v1/surah/number/{number}":
    get:
      tags:
        - "Surah"
      description: "Get Surah by number"
      operationId: "getSurahByNumber"
      parameters:
        - in: path
          name: number
          type: integer
          required: true
          description: Surah number
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Surah"
          schema:
            $ref: "#/definitions/surah"
  "/v1/surah/name/{name}":
    get:
      tags:
        - "Surah"
      description: "Get Surah by name"
      operationId: "getSurahByName"
      parameters:
        - in: path
          name: name
          type: string
          required: true
          description: Surah name
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Surah"
          schema:
            $ref: "#/definitions/surah"
  "/v1/surah/english_name/{englishName}":
    get:
      tags:
        - "Surah"
      description: "Get Surah by english_name"
      operationId: "getSurahByEnglishName"
      parameters:
        - in: path
          name: englishName
          type: string
          required: true
          description: Surah English name
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Surah"
          schema:
            $ref: "#/definitions/surah"
  "/v1/surah/english_name_translation/{englishNameTranslation}":
    get:
      tags:
        - "Surah"
      description: "Get Surah by English name translation"
      operationId: "getSurahByEnglishNameTranslation"
      parameters:
        - in: path
          name: englishNameTranslation
          type: string
          required: true
          description: Surah english name translation
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Surah"
          schema:
            $ref: "#/definitions/surah"
  "/v1/surah/type/{type}":
    get:
      tags:
        - "Surah"
      description: "Get Surah by Revelation Type, Possible values are Meccan or Medinan"
      operationId: "getSurahByRevelationType"
      parameters:
        - in: path
          name: type
          type: string
          required: true
          description: Revelation Type
          enum: [Meccan, Medinan]
        - in: query
          name: limit
          type: integer
          description: Limit the number of results
        - in: query
          name: offset
          type: integer
          description: Start getting result from specific offset usually for next page results
      produces:
        - "application/json"
      responses:
        200:
          description: "Successful operation to get Surah"
          schema:
            $ref: "#/definitions/surah"
...
definitions:
...
  ayah:
    type: object
    properties:
      id:
        type: string
      surahId:
        type: string
      number:
        type: integer
      numberInSurah:
        type: integer
      juz:
        type: integer
      manzil:
        type: integer
      ruku:
        type: integer
      hizbQuarter:
        type: integer
      sajda:
        type: boolean
      arabic:
        type: string
...
  surah:
    type: object
    properties:
      id:
        type: string
      number:
        type: integer
      name:
        type: string
      englishName:
        type: string
      englishNameTranslation:
        type: string
      numberOfAyahs:
        type: integer
      revelationType:
        type: string
...
security: []